+++ maia/io/distribution_tree.py
@@ -99,6 +99,37 @@ def compute_zone_distribution(zone, comm, distri_func):
     for node in PT.iter_children_from_predicates(zone, predicate):
       compute_subset_distribution(node, comm, distri_func)
 
+  mark_global_bcds_arrays(zone)
+
+def mark_global_bcds_arrays(zone):
+  """ Create a Descriptor_t node to indicate the BCDataSet_t/BCData_t/DataArray_t that are
+  global (no pointwise value). This descriptor is stored in the DataSet distribution (if existing)
+  or in BC distribution otherwise.
+  #Size and #Distribution nodes must exist
+  """
+  for bc in PT.get_nodes_from_predicates(zone, 'ZoneBC_t/BC_t'):
+    bc_global_arrays = []
+    
+    for bcds in PT.get_children_from_label(bc, 'BCDataSet_t'):
+      dataset_global_arrays = []
+      for bcdata in PT.get_children_from_label(bcds, 'BCData_t'):
+        is_global_data = lambda n : PT.get_label(n) == 'DataArray_t' \
+                                    and not PT.get_name(n).endswith('#Size') \
+                                    and PT.get_child_from_name(bcdata, PT.get_name(n)+'#Size') is None
+        for data_array in PT.get_children_from_predicate(bcdata, is_global_data):
+          dataset_global_arrays.append(f"{PT.get_name(bcdata)}/{PT.get_name(data_array)}")
+
+      if len(dataset_global_arrays) > 0:
+        distri_bcds_n = MT.getDistribution(bcds)
+        if distri_bcds_n is not None: # Register in BCDS/Distribution node
+          PT.new_Descriptor('BCDataGlobal', '\n'.join(dataset_global_arrays), parent=distri_bcds_n)
+        else: # Save for later registration in BC
+          bc_global_arrays.extend([f'{PT.get_name(bcds)}/{path}' for path in dataset_global_arrays])
+
+    if len(bc_global_arrays) > 0: # Register in BC/Distribution node
+      PT.new_Descriptor('BCDataGlobal', '\n'.join(bc_global_arrays), parent=MT.getDistribution(bc))
+
+
 
 def add_distribution_info(dist_tree, comm, distribution_policy='uniform'):
   """